% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.r
\name{compatible}
\alias{compatible}
\title{Returns whether two clades are compatible}
\usage{
compatible(x, y)
}
\arguments{
\item{x}{a clade.}

\item{y}{a clade.}
}
\value{
logical. \code{TRUE} if \code{x} and \code{y} are compatible and \code{FALSE} otherwise.
}
\description{
Returns whether two clades are compatible
}
\examples{
compatible(x = "{t1,t2}", y = "{t1,t3,t4}")
compatible(x = "{t1,t2}", y = "{t1,t2,t4}")
}
