% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.r
\name{get_clade_elements}
\alias{get_clade_elements}
\title{Get elements of a clade}
\usage{
get_clade_elements(x)
}
\arguments{
\item{x}{a string containing a clade.}
}
\value{
a vector of strings with the elements of \code{x}.
}
\description{
Get elements of a clade
}
\examples{
all.c4 <- make_all_clades(4)
get_clade_elements(all.c4)
}
