% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.r
\name{get_clade_size}
\alias{get_clade_size}
\title{Returns the size of a clade.}
\usage{
get_clade_size(x)
}
\arguments{
\item{x}{a string containing a clade.}
}
\value{
a numeric value with the size of x
}
\description{
Returns the size of a clade.
}
\examples{
all.c4 <- make_all_clades(4)
get_clade_size(all.c4)
}
