% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclades.R
\name{Gn}
\alias{Gn}
\title{Auxiliary function for computing joint probabilities.}
\usage{
Gn(a, b, n, log = FALSE)
}
\arguments{
\item{a}{size of clade A.}

\item{b}{size of clade B.}

\item{n}{number of taxa.}

\item{log}{logical. If \code{TRUE}, returns the log joint probability (default is \code{FALSE}).}
}
\value{
(log) joint probability
}
\description{
Auxiliary function for computing joint probabilities.
}
