% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_prob_clades.R
\name{joint_prob_clades}
\alias{joint_prob_clades}
\title{Computes the joint probability for two clades for a given number of taxa.}
\usage{
joint_prob_clades(c1, c2, n, log = FALSE, verbose = FALSE)
}
\arguments{
\item{c1}{a string containing a clade.}

\item{c2}{a string containing a clade.}

\item{n}{number of taxa.}

\item{log}{logical. If \code{TRUE} computes the log probability (default is \code{FALSE}).}

\item{verbose}{logical. If \code{TRUE} it prints the case (default is \code{FALSE}).}
}
\value{
(log) joint probability.
}
\description{
Computes the joint probability for two clades for a given number of taxa.
}
\examples{
joint_prob_clades(c1 = "{t1,t2}", c2 = "{t1,t2}", n = 5) ## case 1
joint_prob_clades(c1 = "{t1,t2}", c2 = "{t1,t2,t4}", n = 5) ## case 2/3
joint_prob_clades(c1 = "{t1,t2}", c2 = "{t3,t4,t5}", n = 5) ## case 4
joint_prob_clades(c1 = "{t1,t2}", c2 = "{t3,t4}", n = 5) ## case 5
joint_prob_clades(c1 = "{t1,t2}", c2 = "{t1,t3,t4}", n = 5) ## case 6
}
