% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.r
\name{log_diff_exp}
\alias{log_diff_exp}
\title{Logarithm of the difference between two logarithms}
\usage{
log_diff_exp(x, y)
}
\arguments{
\item{x}{log of first value.}

\item{y}{log of second value.}
}
\value{
log of the difference of exp(x) and exp(y).
}
\description{
Logarithm of the difference between two logarithms
}
