% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_building_functions.r
\name{make_all_clades}
\alias{make_all_clades}
\title{Create all (non-trivial) clades on \code{n} taxa.}
\usage{
make_all_clades(n)
}
\arguments{
\item{n}{number of taxa.}
}
\value{
a vector of strings containing all clades on \code{n} taxa.
}
\description{
Creates all clades on \code{n} taxa in the form of a vector of  strings.
Clades are of the form \code{"{t1,t4}"}, \code{"{t1,t2,t4}"}.
}
\details{
Excludes 'trivial' clades such as \code{"{t1}"} or \code{"{t1,t2,t3,t4,t5}"} (for 5 taxa).
}
\examples{
make_all_clades(4)
}
