% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_building_functions.r
\name{make_clade_corr_grid}
\alias{make_clade_corr_grid}
\title{Construct a grid of 'all vs all' correlations for a list of K clades on \code{n} taxa.}
\usage{
make_clade_corr_grid(Clades, n, diagonal = FALSE, ncores = 2)
}
\arguments{
\item{Clades}{a character vector with all clades to be included.}

\item{n}{number of taxa.}

\item{diagonal}{logical. If \code{TRUE} includes the diagonal entries,
which should all be 1 (default is \code{FALSE}).}

\item{ncores}{number of cores to be used.}
}
\value{
a grid of correlations between all clades.
}
\description{
Construct a grid of 'all vs all' correlations for a list of K clades on \code{n} taxa.
}
\examples{
clades.n4 <- make_all_clades(4)
make_clade_corr_grid(clades.n4, n = 4)
}
