% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_building_functions.r
\name{make_clade_corr_mat}
\alias{make_clade_corr_mat}
\title{Build the correlation matrix for (non-trivial) clades on \code{n} taxa.}
\usage{
make_clade_corr_mat(n)
}
\arguments{
\item{n}{number of taxa.}
}
\value{
a  2^n-n-2 x  2^n-n-2 correlation matrix.
}
\description{
Build the correlation matrix for (non-trivial) clades on \code{n} taxa.
}
\examples{
make_clade_corr_mat(4)
}
