% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_building_functions.r
\name{make_clade_joint_mat}
\alias{make_clade_joint_mat}
\title{Build the matrix of joint probabilities for a list of K clades on \code{n} taxa.}
\usage{
make_clade_joint_mat(Clades, n)
}
\arguments{
\item{Clades}{a character vector with all K clades to be included.}

\item{n}{number of taxa.}
}
\value{
a  K x  K matrix of joint probabilities.
}
\description{
Build the matrix of joint probabilities for a list of K clades on \code{n} taxa.
}
\examples{
clades.n4 <- make_all_clades(4)
make_clade_joint_mat(Clades = clades.n4, n = 4)
}
