% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_corr.R
\name{max_corr}
\alias{max_corr}
\title{Maximum possible correlation between clades on \code{n} taxa.}
\usage{
max_corr(n)
}
\arguments{
\item{n}{number of taxa.}
}
\value{
maximum value of the correlation matrix on \code{n} taxa.
}
\description{
Maximum possible correlation between clades on \code{n} taxa.
}
\examples{
ns <- round(seq(4, 1E3, length.out = 200))
plot(ns, max_corr(ns),  lwd = 3,
xlab = expression(n), ylab = expression(rho[max](n)),
 main = "Maximum correlation between clade indicators", 
 type = "l" , cex.lab = 1.5)
 abline(h = 0, lwd = 2, lty = 2)
}
