% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_corr.R
\name{min_corr}
\alias{min_corr}
\title{Minimum possible correlation between clades on \code{n} taxa.}
\usage{
min_corr(n)
}
\arguments{
\item{n}{number of taxa.}
}
\value{
mimimum value of the correlation matrix on \code{n} taxa.
}
\description{
Minimum possible correlation between clades on \code{n} taxa.
}
\examples{
ns <- round(seq(4, 1E3, length.out = 200))
plot(ns, min_corr(ns),  lwd = 3,
xlab = expression(n), ylab = expression(rho[min](n)),
 main = "Minimum correlation between clade indicators", 
 type = "l" , cex.lab = 1.5)
 abline(h = 0, lwd = 2, lty = 2)
}
