% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclades.R
\name{pclades}
\alias{pclades}
\title{Joint probability of two clades}
\usage{
pclades(a, b, n, case = 1, log = FALSE)
}
\arguments{
\item{a}{size of clade A.}

\item{b}{size of clade B.}

\item{n}{number of taxa.}

\item{case}{see details}

\item{log}{logical if TRUE (default is FALSE) returns log probability}
}
\value{
(log) joint probability of clades A and B given their sizes.
}
\description{
Joint probability of two clades
}
\details{
case 1: A == B;
case 2: A is contained in B;
case 3: B is contained in A;
case 4: A and B are disjoint and a + b = n;
case 5: A and B are disjoint and a + b < n;
case 6: A and B are incompatible.
}
\examples{
pclades(a = 2,  b = 3, n = 5,  case = 2)
}
