% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclade.R
\name{prob_clade}
\alias{prob_clade}
\title{Probability of a clade under a PDA model on \code{n} taxa.}
\usage{
prob_clade(clade, n, log = FALSE)
}
\arguments{
\item{clade}{a string containing a clade.}

\item{n}{number of taxa.}

\item{log}{logical. If \code{TRUE} returns the log probability (default is \code{FALSE}).}
}
\value{
(log) probability
}
\description{
Probability of a clade under a PDA model on \code{n} taxa.
}
\examples{
all.c4 <- make_all_clades(4)
prob_clade(all.c4, n = 4)
}
\seealso{
\code{\link[cladeCorrelation]{pclade}}
}
