% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclades.R
\name{Rn}
\alias{Rn}
\title{Auxiliary function for computing joint clade probability.}
\usage{
Rn(a, b, n, log = FALSE)
}
\arguments{
\item{a}{size of first clade.}

\item{b}{size of second clade.}

\item{n}{number of taxa.}

\item{log}{logical. If \code{TRUE}, returns the log (default is \code{FALSE}).}
}
\value{
(log) joint probability for cases 2 and 3.
}
\description{
Auxiliary function for computing joint clade probability.
}
