% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_building_functions.r
\name{make_clade_corr_grid}
\alias{make_clade_corr_grid}
\title{Construct a grid of all (non-trivial) clades on \code{n} taxa.}
\usage{
make_clade_corr_grid(n, diagonal = FALSE, ncores = 2)
}
\arguments{
\item{n}{number of taxa}

\item{diagonal}{logical. If \code{TRUE} includes the diagonal entries,
which should all be 1 (default is \code{FALSE}).}

\item{ncores}{number of cores to be used.}
}
\value{
a grid of correlations between all clades.
}
\description{
Construct a grid of all (non-trivial) clades on \code{n} taxa.
}
\examples{
make_clade_corr_grid(4)
}
