% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_building_functions.r
\name{make_clade_corr_mat}
\alias{make_clade_corr_mat}
\title{Build the correlation matrix for a list of K clades on \code{n} taxa.}
\usage{
make_clade_corr_mat(Clades, n, ncores = 2)
}
\arguments{
\item{Clades}{a character vector with all clades to be included.}

\item{n}{number of taxa.}

\item{ncores}{number of cores to be used.}
}
\value{
a  K x  K correlation matrix.
}
\description{
Build the correlation matrix for a list of K clades on \code{n} taxa.
}
\examples{
clades.n4 <- make_all_clades(4)
make_clade_corr_mat(Clades = clades.n4, n = 4)
}
