% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_building_functions.r
\name{make_clade_joint_mat}
\alias{make_clade_joint_mat}
\title{Build the matrix of joint probabilities for (non-trivial) clades on \code{n} taxa.}
\usage{
make_clade_joint_mat(n)
}
\arguments{
\item{n}{number of}
}
\value{
a  2^n-n-2 x  2^n-n-2 matrix of joint probabilities.
}
\description{
Build the matrix of joint probabilities for (non-trivial) clades on \code{n} taxa.
}
\examples{
make_clade_joint_mat(4)
}
