% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_compat.R
\name{prop_compat}
\alias{prop_compat}
\title{Proportion of clades which are pairwise compatible.}
\usage{
prop_compat(n, trivial = FALSE, log = FALSE)
}
\arguments{
\item{n}{number of taxa.}

\item{trivial}{logical. If \code{TRUE}, trivial clades with 1 and
n members are considered (default is \code{FALSE}).}

\item{log}{logical. \code{TRUE}, the log proportion is returned (default is \code{FALSE}).}
}
\value{
(log) proportion of clades which are pairwise compatible.
}
\description{
Proportion of clades which are pairwise compatible.
}
\examples{
prop_compat(5)
prop_compat(50)
prop_compat(500)
ns <- round(seq(4, 1E3, length.out = 200))
plot(ns, prop_compat(ns), lwd = 3, xlab = expression(n),
ylab = expression(p[c](n)), main = "Proportion compatible", type = "l", cex.lab = 1.5)
}
